[GetUp] First version without rewarding effort
The scene is set as a ragdoll prefab trying to get up from lying on the ground. It is able to lift its hips and legs to minimize negative reward, but still doesn't lift its head. Maybe more training could solve it, but an effort-based reward system could improve this.